start pushing messaging to the periphery
