+++ pipeline/relation_embedder/batcher/src/main/scala/weco/pipeline/batcher/Downstream.scala
@@ -1,0 +1,7 @@
+package weco.pipeline.batcher
+
+import scala.util.Try
+
+trait Downstream {
+  def notify(batch: Batch): Try[Unit]
+}
